Create README.md
Small program I wrote in Matlab attempting to model the equation for single-shot probability of kill from the AIAA Education Series book "Mathematical Methods in Defense Analysis, 3rd Edition" by J.S Przemieniecki. The program prompts the user for two inputs: REP (range error probable) as well as target length. As in the text, the computation assumes no systemic error, and that the aiming point is the center of the target. Additionally, the LEP (linear error probable) in range is assumed to be 0.6745.